Clean up undo buffer allocator, move to separate file
