+++ src/include/duckdb/transaction/duck_transaction_manager.hpp
@@ -14,6 +14,7 @@
 
 namespace duckdb {
 class DuckTransaction;
+struct UndoBufferProperties;
 
 //! The Transaction Manager is responsible for creating and managing
 //! transactions
