+++ src/include/duckdb/transaction/transaction.hpp
@@ -10,10 +10,9 @@
 
 #include "duckdb/catalog/catalog_entry/sequence_catalog_entry.hpp"
 #include "duckdb/common/types/data_chunk.hpp"
-#include "duckdb/transaction/undo_buffer.hpp"
-#include "duckdb/common/atomic.hpp"
 #include "duckdb/transaction/transaction_data.hpp"
 #include "duckdb/common/shared_ptr.hpp"
+#include "duckdb/common/atomic.hpp"
 
 namespace duckdb {
 class SequenceCatalogEntry;
