+++ src/transaction/CMakeLists.txt
@@ -4,6 +4,7 @@ add_library_unity(
   duck_transaction_manager.cpp
   duck_transaction.cpp
   meta_transaction.cpp
+  undo_buffer_allocator.cpp
   undo_buffer.cpp
   transaction_context.cpp
   transaction.cpp
